Connect ticket purchase endpoint to frontend; rename files and folders for clarity
