+++ frontend/src/app/page-registry.ts
@@ -1,13 +1,13 @@
 import { type ComponentType } from 'react';
-import { EventListPage } from '../features/events/EventListPage';
-import { EventDetailsPage } from '../features/events/EventDetailsPage';
-import CreateEventPage from '../features/events/CreateEventPage';
-import EditEventPage from '../features/events/EditEventPage';
-import { OrganizerEventsPage } from '../features/events/OrganizerEventsPage';
-import { MockPaymentPage } from '../features/tickets/MockPaymentPage';
-import { MyEventsPage } from '../features/tickets/MyEventsPage';
-import { PurchaseTicketsPage } from '../features/tickets/PurchaseTicketsPage';
-import { TicketDetailsPage } from '../features/tickets/TicketDetailsPage';
+import { EventListPage } from '../features/event/EventListPage';
+import { EventDetailsPage } from '../features/event/EventDetailsPage';
+import CreateEventPage from '../features/event/CreateEventPage';
+import EditEventPage from '../features/event/EditEventPage';
+import { OrganizerEventsPage } from '../features/event/OrganizerEventsPage';
+import { MockPaymentPage } from '../features/ticket/MockPaymentPage';
+import { MyEventsPage } from '../features/ticket/MyEventsPage';
+import { PurchaseTicketsPage } from '../features/ticket/PurchaseTicketsPage';
+import { TicketDetailsPage } from '../features/ticket/TicketDetailsPage';
 import { roleHomePaths, type UserRole } from './roles';
 
 export type AppPage = {
