+++ frontend/src/features/event/display.ts
