+++ frontend/src/features/ticket/MockQrCode.tsx
